# src/main/resources/api.yaml
openapi: 3.0.0
info:
  description: API specification for SuperBlog microservice
  version: v3
  title: SuperBlog microservice API
servers:
  - url: 'http://localhost:8080/'
paths:
...
  /comments:
    post:
      tags:
        - createComment
      summary: Create Comment
      operationId: createComment
      requestBody:
        description: Create Comment on article
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '200':
          description: Create comment successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectCreationSuccessResponse'
  /comments/activate:
    put:
      summary: Update the status of comment by Id
      description: Update the status of comment by Id
      operationId: UpdateCommentById
      requestBody:
        description: Update Status of comment by id
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRejectedCommentStatusRequest'
      responses:
        '200':
          description: Update Category successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectCreationSuccessResponse'
  /comments/deactivate:
    put:
      summary: Update the status of comment by Id
      description: Update the status of comment by Id
      operationId: UpdateCommentStatusById
      requestBody:
        description: Update Status of comment by id
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAcceptedCommentStatusRequest'
      responses:
        '200':
          description: Update Category successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectCreationSuccessResponse'
components:
  schemas:
...
    UpdateRejectedCommentStatusRequest:
      type: object
      properties:
        commentId:
          type: string
    UpdateAcceptedCommentStatusRequest:
      type: object
      properties:
        commentId:
          type: string
    CreateCommentRequest:
      type: object
      properties:
        content:
          type: string
        articleId:
          type: string
        userAuthorId:
          type: string
    ObjectCreationSuccessResponse:
      type: object
      properties:
        id:
          type: string
        responseCode:
          type: integer
...